pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[project]
name = "toolforgeio"
version = "0.0.6"
description = "Library for building tools in the ToolForge Data Science tools ecosystem using Python"
readme = "README.md"
authors = [{ name = "Andy Boothe", email = "andy@toolforge.io" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["toolforge", "toolforge.io"]
# Some day, we should be able to sync requirements.txt to pyproject.toml...
dependencies = [
    "requests~=2.28.2",
    "openpyxl~=3.0.10",
    "pandas~=1.5.3",
    "PyYAML~=6.0",
    "xlrd~=2.0.1",
    "chardet~=5.1.0",
    'tomli; python_version < "3.11"'
]
requires-python = ">=3.9"

[project.urls]
Homepage = "https://github.com/toolforgeio/toolforge4py"
